Allow admins to fail uploads and move purchase to production
If a customer is unable to complete their upload, this will allow us to
move the purchase into the readyForProduction state.

Also added a progress bar along with the file counts. I think this is
valuable because sometimes a few files are the majority of the upload,
so this gives an indication of how much of the content (bytes) has been
uploaded.