Allow specifying a purchase was shipped with another
We will copy the tracking number and label info to the purchase, and
save the other purchase id on the purchase.

Also changed the code to update a purchase to only update specified
fields to avoid clobbering data.

Removed the archive button that just updates status. We have been
archiving by moving the purchase so it doesn't show up in the app
anymore.